Improved the log output
Lidarr no longer allows the use of parameters, so that put a wrench in how the script determines which folders to scan and convert. The script has been modified to allow for no parameters to be passed, parse folders dropped on the compiled EXE, or as a standalone app. Change the variables at the start of the script to your own locations and its good to go. 

Change MusicFolder to your library location
Change BitRate to your desired output rate
Change YesLog to enable or disable log output
Change YesDelete to enable or disable deleting the old file (most likely required now)
Change ShowGui to enable or disable the script being a standalone app